refactor(tests): move registration capacity testing to own file
