+++ events/tests/test_event_get.py
@@ -8,7 +8,6 @@ from django.conf import settings
 from django.contrib.gis.gdal import CoordTransform, SpatialReference
 from django.contrib.gis.geos import Point
 from django.db import DEFAULT_DB_ALIAS, connections
-from django.test import TestCase
 from django.test.utils import CaptureQueriesContext
 from django.utils import timezone
 from django.utils.timezone import localtime
@@ -1961,76 +1960,3 @@ def test_event_list_shows_image_and_its_license(event, image):
     response = get_list(api_client)
     license_url = response.data["data"][0]["images"][0]["license_url"]
     assert license_url == "http://urltothe.license"
-
-
-class FilterEventsByRegistrationCapacitiesV1TestCase(TestCase, EventsListTestCaseMixin):
-    @classmethod
-    def setUpTestData(cls):
-        cls.list_url = reverse("event-list", version="v1")
-
-    def test_get_event_list_registration_remaining_attendee_capacity_gte(self):
-        registration = RegistrationFactory(remaining_attendee_capacity=10)
-        registration2 = RegistrationFactory(remaining_attendee_capacity=5)
-        registration3 = RegistrationFactory(remaining_attendee_capacity=1)
-
-        for capacity, events in [
-            (1, [registration.event, registration2.event, registration3.event]),
-            (5, [registration.event, registration2.event]),
-            (10, [registration.event]),
-            (11, []),
-        ]:
-            with self.subTest():
-                self._get_list_and_assert_events(
-                    f"registration__remaining_attendee_capacity__gte={capacity}", events
-                )
-
-    def test_get_event_list_registration_remaining_waiting_list_capacity_gte(self):
-        registration = RegistrationFactory(remaining_waiting_list_capacity=10)
-        registration2 = RegistrationFactory(remaining_waiting_list_capacity=5)
-        registration3 = RegistrationFactory(remaining_waiting_list_capacity=1)
-
-        for capacity, events in [
-            (1, [registration.event, registration2.event, registration3.event]),
-            (5, [registration.event, registration2.event]),
-            (10, [registration.event]),
-            (11, []),
-        ]:
-            with self.subTest():
-                self._get_list_and_assert_events(
-                    f"registration__remaining_waiting_list_capacity__gte={capacity}",
-                    events,
-                )
-
-    def test_get_event_list_registration_remaining_attendee_capacity_isnull(self):
-        registration = RegistrationFactory(remaining_attendee_capacity=10)
-        registration2 = RegistrationFactory(remaining_attendee_capacity=None)
-        registration3 = RegistrationFactory(remaining_attendee_capacity=1)
-
-        for capacity, events in [
-            (1, [registration2.event]),
-            (0, [registration.event, registration3.event]),
-            (True, [registration2.event]),
-            (False, [registration.event, registration3.event]),
-        ]:
-            with self.subTest():
-                self._get_list_and_assert_events(
-                    f"registration__remaining_attendee_capacity__isnull={capacity}",
-                    events,
-                )
-
-    def test_get_event_list_registration_remaining_waiting_list_capacity_isnull(self):
-        registration = RegistrationFactory(remaining_waiting_list_capacity=10)
-        registration2 = RegistrationFactory(remaining_waiting_list_capacity=None)
-        registration3 = RegistrationFactory(remaining_waiting_list_capacity=1)
-
-        for capacity, events in [
-            (1, [registration2.event]),
-            (0, [registration.event, registration3.event]),
-            (True, [registration2.event]),
-            (False, [registration.event, registration3.event]),
-        ]:
-            with self.subTest():
-                self._get_list_and_assert_events(
-                    f"registration__remaining_waiting_list_capacity__isnull={capacity}",
-                    events,
-                )
